requirements.txt
appdirs==1.4.4
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.1.0
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16
Cython==0.29.33
docker-pycreds==0.4.0
filelock==3.9.0
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub==0.13.1
idna==3.4
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
pandas==1.5.3
pathtools==0.1.2
Pillow==9.4.0
protobuf==4.22.1
psutil==5.9.4
pycocotools==2.0.6
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
sentry-sdk==1.16.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
tokenizers==0.13.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
transformers==4.26.1
typing_extensions==4.5.0
urllib3==1.26.14
wandb==0.13.11
wcwidth==0.2.6
